% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_PlasmoSim_loaded}
\alias{check_PlasmoSim_loaded}
\title{Check that PlasmoSim package has loaded successfully}
\usage{
check_PlasmoSim_loaded()
}
\description{
Simple function to check that PlasmoSim package has loaded 
  successfully. Prints "PlasmoSim loaded successfully!" if so.
}
