% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_haplotype_identity}
\alias{get_haplotype_identity}
\title{Get proportion identical between two haplotype matrices}
\usage{
get_haplotype_identity(mat1, mat2)
}
\arguments{
\item{mat1, mat2}{matrices representing sets of haplotypes to compare.
Haplotypes are in rows and loci are in columns.}
}
\description{
Compare two sets of haplotypes (matrices), and return the
  proportion of identical sites over all pairwise comparisons. Values can be
  any numeric value; for example if values represent ancestry then this
  function returns the average identity by descent, or if values represent
  alleles then it returns the average identity by state.
}
