% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_identity_matrix}
\alias{get_identity_matrix}
\title{Get pairwise genetic identity matrix}
\usage{
get_identity_matrix(sim_output, deme_level = FALSE)
}
\arguments{
\item{sim_output}{simulation output from \code{sim_falciparum()}.}

\item{deme_level}{if \code{TRUE} then return pairwise identity at the deme
level, averaged over all individuals within each deme. Otherwise return at
the individual level.}
}
\description{
Calculates pairwise genetic identity between all samples. If
  \code{deme_level = TRUE} this is averaged over all individuals within a
  deme, to produce average pairwise relatedness within and between demes.
  Each time point in the sample is considered independently, and output as a
  list.
}
