% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_spatial_distance}
\alias{get_spatial_distance}
\title{Get great circle distance between spatial points}
\usage{
get_spatial_distance(lat, lon)
}
\arguments{
\item{lat}{vector of latitudes.}

\item{lon}{vector of longitudes.}
}
\description{
Get great circle distance between spatial points.
}
