% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlasmoSim.R
\docType{package}
\name{goodegg}
\alias{goodegg}
\title{PlasmoSim}
\description{
A basic Plasmodium simulator. Contains functions to simulate
  epidemiological and genetic data from a simple model of Plasmodium
  falciparum transmission.
}
